Create BubbleSort.java
Can try to make bubbleSortV more efficient. If it checks each pair in the list and no swapping happens during that one pass (meaning the list is sorted already), then that should be the only pass it would make (so only 1 pass is needed).

So if we, by chance, have a 100-sized list that is completely sorted, we don't need to do 99 passes, when you only need 1 to check if it's all sorted.